open-ai-create_embeddings/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
frozenlist==1.3.3
idna==3.4
joblib==1.2.0
loguru==0.7.0
lxml==4.9.2
multidict==6.0.4
nltk==3.8.1
numpy==1.24.3
openai==0.27.7
python-dateutil==2.8.2
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
site-map-parser==0.3.9
six==1.16.0
soupsieve==2.4.1
tqdm==4.65.0
typing_extensions==4.6.1
urllib3==2.0.2
yarl==1.9.2
psycopg2-binary==2.9.1
pgvector